Display object name instead of the tag slug
